# csi-back/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "ML Service",
      "version": "0.1.0"
    },
    "paths": {
...
      "/api/v1/html-analyze/extract-text": {
        "post": {
          "tags": [
            "HTML 静态分析"
          ],
          "summary": "提取 HTML 中的纯文本",
          "operationId": "extract_text_api_v1_html_analyze_extract_text_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HtmlExtractTextRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HtmlExtractTextResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/html-analyze/clean": {
        "post": {
          "tags": [
            "HTML 静态分析"
          ],
          "summary": "清理 HTML 标签与脚本",
          "operationId": "clean_html_api_v1_html_analyze_clean_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HtmlCleanRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HtmlCleanResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/html-analyze/extract-links": {
        "post": {
          "tags": [
            "HTML 静态分析"
          ],
          "summary": "提取 HTML 中所有资源链接",
          "operationId": "extract_links_api_v1_html_analyze_extract_links_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HtmlExtractLinksRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HtmlExtractLinksResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/html-analyze/extract-text-batch": {
        "post": {
          "tags": [
            "HTML 静态分析"
          ],
          "summary": "批量提取 HTML 中的纯文本",
          "operationId": "extract_text_batch_api_v1_html_analyze_extract_text_batch_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HtmlBatchRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/HtmlExtractTextBatchItem"
                    },
                    "type": "array",
                    "title": "Response Extract Text Batch Api V1 Html Analyze Extract Text Batch Post"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/html-analyze/clean-batch": {
        "post": {
          "tags": [
            "HTML 静态分析"
          ],
          "summary": "批量清理 HTML 标签与脚本",
          "operationId": "clean_html_batch_api_v1_html_analyze_clean_batch_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HtmlBatchRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/HtmlCleanBatchItem"
                    },
                    "type": "array",
                    "title": "Response Clean Html Batch Api V1 Html Analyze Clean Batch Post"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/html-analyze/extract-links-batch": {
        "post": {
          "tags": [
            "HTML 静态分析"
          ],
          "summary": "批量提取 HTML 中所有资源链接",
          "operationId": "extract_links_batch_api_v1_html_analyze_extract_links_batch_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HtmlBatchRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/HtmlExtractLinksBatchItem"
                    },
                    "type": "array",
                    "title": "Response Extract Links Batch Api V1 Html Analyze Extract Links Batch Post"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
        "HtmlBatchItem": {
          "properties": {
            "uuid": {
              "type": "string",
              "title": "Uuid",
              "description": "本条数据的唯一标识"
            },
            "html": {
              "type": "string",
              "title": "Html",
              "description": "HTML 内容"
            }
          },
          "type": "object",
          "required": [
            "uuid",
            "html"
          ],
          "title": "HtmlBatchItem"
        },
        "HtmlBatchRequest": {
          "properties": {
            "datas": {
              "items": {
                "$ref": "#/components/schemas/HtmlBatchItem"
              },
              "type": "array",
              "minItems": 1,
              "title": "Datas",
              "description": "待处理的数据列表"
            }
          },
          "type": "object",
          "required": [
            "datas"
          ],
          "title": "HtmlBatchRequest"
        },
        "HtmlCleanBatchItem": {
          "properties": {
            "uuid": {
              "type": "string",
              "title": "Uuid",
              "description": "本条数据的唯一标识"
            },
            "html": {
              "type": "string",
              "title": "Html",
              "description": "清理后的 HTML"
            }
          },
          "type": "object",
          "required": [
            "uuid",
            "html"
          ],
          "title": "HtmlCleanBatchItem"
        },
        "HtmlCleanRequest": {
          "properties": {
            "html": {
              "type": "string",
              "title": "Html",
              "description": "待清理的 HTML 内容"
            }
          },
          "type": "object",
          "required": [
            "html"
          ],
          "title": "HtmlCleanRequest"
        },
        "HtmlCleanResponse": {
          "properties": {
            "html": {
              "type": "string",
              "title": "Html",
              "description": "清理后的 HTML"
            }
          },
          "type": "object",
          "required": [
            "html"
          ],
          "title": "HtmlCleanResponse"
        },
        "HtmlExtractLinksBatchItem": {
          "properties": {
            "uuid": {
              "type": "string",
              "title": "Uuid",
              "description": "本条数据的唯一标识"
            },
            "links": {
              "items": {
                "type": "string"
              },
              "type": "array",
              "title": "Links",
              "description": "提取出的资源链接列表"
            }
          },
          "type": "object",
          "required": [
            "uuid",
            "links"
          ],
          "title": "HtmlExtractLinksBatchItem"
        },
        "HtmlExtractLinksRequest": {
          "properties": {
            "html": {
              "type": "string",
              "title": "Html",
              "description": "待提取资源链接的 HTML 内容"
            }
          },
          "type": "object",
          "required": [
            "html"
          ],
          "title": "HtmlExtractLinksRequest"
        },
        "HtmlExtractLinksResponse": {
          "properties": {
            "links": {
              "items": {
                "type": "string"
              },
              "type": "array",
              "title": "Links",
              "description": "提取出的资源链接列表"
            }
          },
          "type": "object",
          "required": [
            "links"
          ],
          "title": "HtmlExtractLinksResponse"
        },
        "HtmlExtractTextBatchItem": {
          "properties": {
            "uuid": {
              "type": "string",
              "title": "Uuid",
              "description": "本条数据的唯一标识"
            },
            "text": {
              "type": "string",
              "title": "Text",
              "description": "提取出的纯文本"
            }
          },
          "type": "object",
          "required": [
            "uuid",
            "text"
          ],
          "title": "HtmlExtractTextBatchItem"
        },
        "HtmlExtractTextRequest": {
          "properties": {
            "html": {
              "type": "string",
              "title": "Html",
              "description": "待提取文本的 HTML 内容"
            }
          },
          "type": "object",
          "required": [
            "html"
          ],
          "title": "HtmlExtractTextRequest"
        },
        "HtmlExtractTextResponse": {
          "properties": {
            "text": {
              "type": "string",
              "title": "Text",
              "description": "提取出的纯文本"
            }
          },
          "type": "object",
          "required": [
            "text"
          ],
          "title": "HtmlExtractTextResponse"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      }
    }
  }